back/requirements.txt
# Core Framework
fastapi[standard]>=0.118.0
uvicorn[standard]>=0.31.0

# Database
sqlalchemy>=2.0.43
sqlmodel>=0.0.25

# Authentication & Security
#supabase>=2.21.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Environment & Configuration
python-dotenv>=1.0.0

# Development Tools (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0


#.venv\Scripts\activate
#pip install -r requirements.txt
#uvicorn main:app --reload